requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pandas>=1.3.0
plotly>=5.15.0